4.4.FlaskLogIn/requirements.txt
python-dotenv
flask
Werkzeug
Flask-Mail
Flask-Session
pydantic
bcrypt
psycopg2-binary


google-generativeai

